Adds a Class that "clones" directoried, improvres methods and deletes some redundant classes.
Adds a class that copies all files from a directory to a target and sets
the lastModified time to the same time (lastModified-Long is not
accurate enough, have to set both, at source and target).

Some methods improved.

DirectoryChanges#getModifiedFiles is now #getChangedFiles and only
returns changed files. Method is adjusted to this.

Path#relativeTo returns the unchanged path, if the path does not start
with the given directory the path should be realtive to. (Previously the
trailing sperator was removed everytime).

Created a new method for make a path realtive inside of the
DirectoryChanges class (will do Path#relativeTo two times, once on the
current and once on the previous directory) called #makeRelative.

FileList#getMappedHashedModifications improves: JSON file is ignored by
default and the value is now the complete path, not only the relative.