Adds new classes, improves classes, logging and comments.
Adds new classes, that connects objects, saving Lists to JSON, helps
with SWT GUIs and a class that helps creating JSONCreators.

Improves the Path object by adding a #touch method for creating a File
on a path.
Improves the FileList object: Symbolic Links new will be ignored. Also
implements this into DirectoryChanges and DirectorySync.
Improves the JSONFile object: The JSON data can be set and the File will
be created if it does not exists.
Improves the JSONCreators object: Now the complete object can be
string-compatible as key and value. Also implements this into JSONMap.
Improves the MultiFileGrep object: It now can be used as an object.

Logging improved in DirectorySync: Now logs how many files will be
copied or removed.