Improves JSON handling among others with more supported objects.
Adds the abstract class JSONCollection to help with JSON and
collections. It's an abstraction of the JSONList class.
Adds the new interface JSONable to gurantee JSON support.

Improves JSONFile with simpler file reading and enables setting of the
JSON data.
Implements the JSONList class as a instance of the JSONCollection.
Implements the new class JSONSet to help with JSON and sets as a
instance of JSONCollection.
Implements JSONable into JSONCreator and AbstractJSONCreator.
Implements JSONable into JSON creators in JSONCreators.

Resolves deprication problems.